Crate cld2_sys

Unsafe, low-level wrapper around cld2, the "compact language detector" based on Chromium's code, plus a very thin C wrapper layer. Normally you won't want to use this library directly unless you're writing your own cld2 wrapper library.

If you need access to APIs which are not currently wrapped, please feel free to send pull requests!

Structs

CLDHints

ResultChunk

Enums

Encoding

Language

Statics

NUM_ENCODINGS

NUM_LANGUAGES

kCLDFlagCr

kCLDFlagEcho

kCLDFlagHtml

kCLDFlagQuiet

kCLDFlagScoreAsQuads

kCLDFlagVerbose

Functions

CLD2_DetectLanguage

CLD2_DetectLanguageSummary

CLD2_DetectLanguageSummary2

CLD2_DetectLanguageVersion

CLD2_ExtDetectLanguageSummary

CLD2_ExtDetectLanguageSummary2

CLD2_ExtDetectLanguageSummary3

CLD2_ExtDetectLanguageSummary4

CLD2_GetLanguageFromName

CLD2_LanguageCode

CLD2_LanguageDeclaredName

CLD2_LanguageName

CLD2_ResultChunkVector_data

CLD2_ResultChunkVector_delete

CLD2_ResultChunkVector_new

CLD2_ResultChunkVector_size

Type Definitions

ResultChunks